FIX: numerical types replaced for size_t to try to solve the Issue #2 feat (off): reader reader aims to open files and create a msg text to decode that came from input (argv)

+++ decoder/danielgorithm.c
@@ -5,9 +5,9 @@ static short find_spot(char c1, char c2)
 	return (c1 == ' ' && c2 == ' ');
 }
 
-static void	fill_symbols(unsigned char **symbols, unsigned char *daniel, int spot)
+static void	fill_symbols(unsigned char **symbols, unsigned char *daniel, size_t spot)
 {
-	int i = 0;
+	size_t i = 0;
 
 	*symbols = (unsigned char *)calloc(strlen((char *)&daniel[spot + 2]), sizeof(unsigned char));
 	spot += 2;
@@ -23,9 +23,9 @@ static void	fill_symbols(unsigned char **symbols, unsigned char *daniel, int spo
 static void split_code_into_map(unsigned char *code, unsigned char **map, unsigned char *symbols)
 {
 	unsigned char mapped_code[8];
-	int code_index = 0;
-	int mapped_index = 0;
-	int symbols_index = 0;
+	size_t code_index = 0;
+	size_t mapped_index = 0;
+	size_t symbols_index = 0;
 
 	bzero(mapped_code, 8);
 	while (code[code_index])
@@ -46,9 +46,9 @@ static void split_code_into_map(unsigned char *code, unsigned char **map, unsign
 	}
 }
 
-static void	fill_map_from_daniel(t_map map, unsigned char *daniel, int spot, unsigned char *symbols)
+static void	fill_map_from_daniel(t_map map, unsigned char *daniel, size_t spot, unsigned char *symbols)
 {
-	int i = 0; 
+	size_t i = 0; 
 
 	unsigned char code[spot + 1];
 	while (i < spot)
@@ -63,7 +63,7 @@ static void	fill_map_from_daniel(t_map map, unsigned char *daniel, int spot, uns
 t_daniel	*danielgorithm(unsigned char *daniel)
 {
 	t_daniel	*new_dani;
-	int i = 0;
+	size_t i = 0;
 
 	new_dani = (t_daniel *)calloc(1, sizeof(t_daniel));
 	while (daniel[i])
